Please enter the commit message for your changes. Lines starting  with '' will be ignored, and an empty message aborts the commit.
 On branch master
 Your branch is up to date with 'origin/master'.

 Changes to be committed:
	modified:   app/Http/Controllers/Admin/MovieController.php
	modified:   app/Services/MovieService.php
	new file:   database/migrations/2024_03_25_152950_add_vietsub_column_to_movies_table.php
	new file:   public/uploads/movie/161816069_260413892235458_2962913939142159672_n-268.jpg
	new file:   public/uploads/movie/161816069_260413892235458_2962913939142159672_n-2843.jpg
	new file:   public/uploads/movie/161816069_260413892235458_2962913939142159672_n-3174.jpg
	new file:   public/uploads/movie/161816069_260413892235458_2962913939142159672_n-3529.jpg
	new file:   public/uploads/movie/161816069_260413892235458_2962913939142159672_n-3800.jpg
	new file:   public/uploads/movie/161816069_260413892235458_2962913939142159672_n-5066.jpg
	new file:   public/uploads/movie/161816069_260413892235458_2962913939142159672_n-5769.jpg
	new file:   public/uploads/movie/161816069_260413892235458_2962913939142159672_n-690.jpg
	new file:   public/uploads/movie/161816069_260413892235458_2962913939142159672_n-7099.jpg
	new file:   public/uploads/movie/161816069_260413892235458_2962913939142159672_n-9462.jpg
	new file:   public/uploads/movie/89970411_806080629888603_1887854386785288192_n-4159.jpg
	modified:   resources/views/admin/movie/form.blade.php
	modified:   resources/views/admin/movie/index.blade.php
	modified:   resources/views/pages/category.blade.php
	modified:   resources/views/pages/country.blade.php
	modified:   resources/views/pages/genre.blade.php
	modified:   resources/views/pages/home.blade.php
	modified:   resources/views/pages/movie.blade.php

+++ app/Http/Controllers/Admin/MovieController.php
@@ -62,25 +62,19 @@ class MovieController extends Controller
         $image = $this->movieService->uploadImage($request);
 
         if($image != false){
-            $store = Movie::create([
-                'title'=>$request->title,
-                'description'=>$request->description,
-                'status' => $request->status,
-                'slug' => $request->slug,
-                'image'=> $image['image'],
-                'country_id' => $request->country_id,
-                'name_eng' => $request->name_eng,
-                'resolution'=>$request->resolution,
-                
-                'genre_id' => $request->genre_id,
-                'hot_movie' => $request->hot_movie,
+            $create = $this->movieService->create($request,$image);
+            if($create){
 
-                'category_id' => $request->category_id
-
+                return redirect()->back();
+            } 
+            return response()->json([
+                'error'=>'Lỗi không tạo được'
             ]);
-            return redirect()->back();
+            
         }
-        return false;
+        return response()->json([
+            'error'=>'Lỗi không xác định'
+        ]);
 
     }
 
@@ -132,28 +126,25 @@ class MovieController extends Controller
         $image = $this->movieService->uploadImage($request);
         $movie = Movie::find($id);
         $get_image = $request->file('image');
-        if(!empty($movie->image) && isset($request->image)){
-            unlink('uploads/movie/'.$movie->image);
+        $file_exist = $this->movieService->file_exist($movie);
+        if(!$file_exist){
+            return response()->json([
+                'msg' => 'Loi khong xoa duoc file'
+            ]);
         }
         if($image != false){
-            $movie->update([
-                'title'=>$request->title,
-                'description'=>$request->description,
-                'status' => $request->status,
-                'slug' => $request->slug,
-                'name_eng' => $request->name_eng,
-                'image'=> $image['image'] ?? null,
-                'country_id' => $request->country_id,
-                'genre_id' => $request->genre_id,
-                'resolution'=>$request->resolution,
+            $update = $this->movieService->update($request, $movie,$image);
+            if($update){
 
-                'category_id' => $request->category_id,
-                'hot_movie' => $request->hot_movie,
-
+                return redirect()->back();
+            }
+            return response()->json([
+                'error'=>'Lỗi không cat nhat duoc'
             ]);
-            return redirect()->back();
         }
-        // return false;
+        return response()->json([
+            'error'=>'Lỗi không xác định'
+        ]);
     }
 
     /**
@@ -165,8 +156,11 @@ class MovieController extends Controller
     public function destroy($id)
     {
         $movie =Movie::find($id);
-        if(!empty($movie->image)){
-            unlink('uploads/movie/'.$movie->image);
+        $file_exist = $this->movieService->file_exist($movie);
+        if(!$file_exist){
+            return response()->json([
+                'msg' => 'Loi khong xoa duoc file'
+            ]);
         }
         $movie->delete();
         return redirect()->back();
